requirements.txt
# Core framework
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.0
pydantic-settings>=2.0

# Database
sqlalchemy[asyncio]>=2.0
aiosqlite>=0.20.0

# Auth
python-jose[cryptography]>=3.3.0
cryptography>=43.0

# HTTP client (agent -> marketplace)
httpx>=0.27.0

# Utilities
python-dotenv>=1.0
python-multipart>=0.0.9

# Agent framework (Phase 3)
# google-adk>=1.24.0
# a2a-sdk[http-server,sql]>=0.3.20

# Payments (Phase 2)
# x402>=2.0.0
